services/vision_plante/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
numpy
minio
psycopg2-binary